backend/requirements.txt
Flask
transformers
groq
Werkzeug
python-dotenv
torch